Implemented the UI layer for the Profile Screen on Android, YT KMM Social Media App episode 18

+++ androidApp/src/main/java/com/dipumba/ytsocialapp/android/account/edit/EditProfileViewModel.kt
@@ -20,7 +20,7 @@ class EditProfileViewModel : ViewModel(){
     var bioTextFieldValue: TextFieldValue by mutableStateOf(TextFieldValue(text = ""))
         private set
 
-    fun fetchProfile(userId: Int){
+    fun fetchProfile(userId: Long){
         viewModelScope.launch {
             uiState = uiState.copy(isLoading = true)
 
@@ -28,7 +28,7 @@ class EditProfileViewModel : ViewModel(){
 
             uiState = uiState.copy(
                 isLoading = false,
-                profile = sampleProfiles.find { it.id == userId }
+                profile = sampleProfiles.find { it.id == userId.toInt() }
             )
 
             bioTextFieldValue = bioTextFieldValue.copy(
