+++ androidApp/src/main/java/com/dipumba/ytsocialapp/android/home/Home.kt
@@ -19,7 +19,7 @@ fun Home(
         postsFeedUiState = viewModel.postsFeedUiState,
         homeRefreshState = viewModel.homeRefreshState,
         onUiAction = {viewModel.onUiAction(it)},
-        onProfileNavigation = {navigator.navigate(ProfileDestination(it.toInt()))},
+        onProfileNavigation = {navigator.navigate(ProfileDestination(it))},
         onPostDetailNavigation = {navigator.navigate(PostDetailDestination(it.postId.toString()))}
     )
 }
